mo_agent_dev/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pydantic==1.10.0a1
sqlalchemy==1.4.25
psycopg2-binary==2.9.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==0.19.2
influxdb-client==1.21.0